requirements.txt
easyocr==1.7.2
matplotlib==3.10.3
numpy==2.3.0
Pillow==11.2.1
PyQt5==5.15.11
PyQt5_sip==12.17.0
skimage==0.0